clean up remove unused files update logos refactor component imports
