+++ src/components/elements/index.ts
@@ -1,0 +1,23 @@
+import AccountIcon from './AccountIcon';
+import Button from './Button';
+import DropDown from './DropDown';
+import Input from './Input';
+import Item from './Item';
+import PageTitle from './PageTitle';
+import Spinner from './Spinner';
+import Tab from './Tab';
+import Table from './Table';
+import Tabs from './Tabs';
+
+export {
+  AccountIcon,
+  Button,
+  DropDown,
+  Input,
+  Item,
+  PageTitle,
+  Spinner,
+  Tab,
+  Table,
+  Tabs,
+};
