+++ src/components/layout/PageContainer.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 interface OverrideProps {
   rowLayout?: boolean;
 }
