+++ src/components/layout/index.ts
@@ -1,0 +1,11 @@
+import Footer from './Footer';
+import Header from './Header';
+import PageContainer from './PageContainer';
+import PageHeader from './PageHeader';
+
+export {
+  Footer,
+  Header,
+  PageContainer,
+  PageHeader,
+};
